[refact] 주문 서비스 에러처리 고도화

+++ msa.order/src/main/java/com/msa/order/presentation/request/CreateOrderRequest.java
@@ -13,7 +13,7 @@ public record CreateOrderRequest(
 	String itemName,
 
 	@NotNull
-	int quantity,
+	Long quantity,
 
 	String description,
